--- tugas/service/lib/orm.js
@@ -1,14 +1,0 @@
-const { createConnection } = require('typeorm');
-
-function connect(entities, config) {
-  return createConnection({
-    ...config,
-    synchronize: true,
-    timezone: 'Asia/Jakarta',
-    entities,
-  });
-}
-
-module.exports = {
-  connect,
-};
