+++ tugas/service/lib/orm.ts
@@ -1,0 +1,9 @@
+import { Connection, ConnectionOptions, createConnection, EntitySchema }  from 'typeorm';
+
+export function connect(entities:EntitySchema[], config: ConnectionOptions) {
+  return createConnection({
+    ...config,
+    synchronize: true,
+    entities,
+  });
+}
