+++ tugas/service/nodemon.json
@@ -1,0 +1,6 @@
+{
+    "ignore": ["**/*.test.ts", "**/*.spec.ts", "node_modules"],
+    "watch": ["service"],
+    "exec": "npm run service",
+    "ext": "ts"
+  }
